Fixing an issue with multi acount support. (#1259)
