+++ k8s/scalyr-service-account.yaml
@@ -11,7 +11,7 @@ metadata:
   name: scalyr-clusterrole
 rules:
 - apiGroups: [""]
-  resources: ["namespaces","pods","replicationcontrollers"]
+  resources: ["namespaces","pods","replicationcontrollers", "secrets"]
   verbs: ["get","list"]
 - apiGroups: [""]
   resources: ["nodes"]
