+++ tests/e2e/scalyr-agent-2-daemonset.yaml
@@ -1,3 +1,47 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: scalyr-config-agent-d
+  namespace: scalyr
+data:
+  k8s_logs.json: |-
+    {
+      "k8s_logs": [
+        {
+          "attributes": {
+            "container_name": "${k8s_container_name}"}
+        }
+      ]
+    }
+  api-key.json: |-
+    {
+       "import_vars": [ "SCALYR_API_KEY" ],
+       "api_key": "$SCALYR_API_KEY"
+      }
+  k8s_events.json: |-
+    {
+      "monitors": [
+        {
+          "module": "scalyr_agent.builtin_monitors.kubernetes_events_monitor"
+        }
+      ]
+    }
+  docker.json: |-
+    {
+      "monitors":[
+        {
+          "module": "scalyr_agent.builtin_monitors.kubernetes_monitor",
+          "stop_agent_on_failure": true
+        }
+      ]
+    }
+  scalyr-server.json: |-
+    {
+        "import_vars": [ { "var": "SCALYR_SERVER", "default": "https://agent.scalyr.com" } ],
+         "scalyr_server": "$SCALYR_SERVER"
+    }
+---
+---
 apiVersion: apps/v1
 kind: DaemonSet
 metadata:
@@ -63,6 +107,8 @@ spec:
         terminationMessagePath: /dev/termination-log
         terminationMessagePolicy: File
         volumeMounts:
+        - name: scalyr-config-agent-d-vol
+          mountPath: /etc/scalyr-agent-2/agent.d
         - mountPath: /var/lib/docker/containers
           name: varlibdockercontainers
           readOnly: true
@@ -101,6 +147,9 @@ spec:
       serviceAccountName: scalyr-service-account
       terminationGracePeriodSeconds: 30
       volumes:
+      - name: scalyr-config-agent-d-vol
+        configMap:
+          name: scalyr-config-agent-d
       - hostPath:
           path: /var/lib/docker/containers
           type: ""
